refactored BusinessInfoWindow
