+++ src/components/BusinessInfoWindow/CloseButton.tsx
@@ -1,0 +1,13 @@
+import { FaXmark } from 'react-icons/fa6';
+
+export const CloseButton = ({ onClick }: { onClick: () => void }) => {
+	return (
+		<button
+			type="button"
+			onClick={onClick}
+			className="bg-neutral-300/50 hover:bg-neutral-300/60 transition-all cursor-pointer rounded-full h-[30px] w-[30px] flex items-center justify-center text-neutral-50 border border-neutral-100/25 shadow-lg"
+		>
+			<FaXmark size={16} />
+		</button>
+	);
+};
